pyproject.toml
[tool.poetry]
name = "counterdev"
version = "0.1.0"
description = ""
authors = ["Irae Hueck Costa <mail@irae.me>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
